+++ src/gui/ldArc.h
@@ -28,14 +28,14 @@ typedef struct {
     bool isClockwise:1;
     ldColor color[2];
     ldColor parentColor;
-    uint32_t srcAddr;
-    uint32_t maskAddr;
+    uintptr_t srcAddr;
+    uintptr_t maskAddr;
     uint16_t startAngle_x10[2];
     uint16_t endAngle_x10[2];
     uint16_t rotationAngle_x10;
 }ldArc_t;
 
-ldArc_t* ldArcInit(uint16_t nameId, uint16_t parentNameId, int16_t x, int16_t y, int16_t width, int16_t height, uint32_t srcQuarterAddr, uint32_t maskQuarterAddr, uint16_t parentColor);
+ldArc_t* ldArcInit(uint16_t nameId, uint16_t parentNameId, int16_t x, int16_t y, int16_t width, int16_t height, uintptr_t srcQuarterAddr, uintptr_t maskQuarterAddr, uint16_t parentColor);
 void ldArcFrameUpdate(ldArc_t* pWidget);
 void ldArcLoop(ldArc_t *pWidget,const arm_2d_tile_t *pParentTile,bool bIsNewFrame);
 void ldArcDel(ldArc_t *pWidget);
