+++ src/gui/ldButton.h
@@ -19,9 +19,9 @@ typedef struct {
     bool isWithReleaseMask:1;
     bool isWithPressMask:1;
     bool isCorner:1;
-    uint32_t releaseImgAddr;
-    uint32_t pressImgAddr;
-    uint32_t selectMaskAddr;
+    uintptr_t releaseImgAddr;
+    uintptr_t pressImgAddr;
+    uintptr_t selectMaskAddr;
     uint32_t keyValue;
     ldFontDict_t* pFontDict;
     uint8_t* pStr;
@@ -38,8 +38,8 @@ void ldButtonLoop(ldButton_t *pWidget,const arm_2d_tile_t *pParentTile,bool bIsN
 void ldButtonDel(ldButton_t *pWidget);
 
 void ldButtonSetColor(ldButton_t* pWidget, ldColor releaseColor, ldColor pressColor);
-void ldButtonSetImage(ldButton_t* pWidget,uint32_t releaseImgAddr,bool isReleaseMask,uint32_t pressImgAddr,bool isPressMask);
-void ldButtonSetSelectImage(ldButton_t* pWidget,uint32_t selectMaskAddr,ldColor selectColor);
+void ldButtonSetImage(ldButton_t* pWidget,uintptr_t releaseImgAddr,bool isReleaseMask,uintptr_t pressImgAddr,bool isPressMask);
+void ldButtonSetSelectImage(ldButton_t* pWidget,uintptr_t selectMaskAddr,ldColor selectColor);
 
 void ldButtonSetTransparent(ldButton_t* pWidget,bool isTransparent);
 void ldButtonSetRoundCorner(ldButton_t* pWidget,bool isCorner);
