+++ src/gui/ldImage.h
@@ -19,14 +19,14 @@ typedef struct {
 #endif
 }ldImage_t;
 
-ldImage_t* ldImageInit(uint16_t nameId, uint16_t parentNameId, int16_t x,int16_t y,int16_t width,int16_t height,uint32_t imageAddr,bool isWithMask);
+ldImage_t* ldImageInit(uint16_t nameId, uint16_t parentNameId, int16_t x,int16_t y,int16_t width,int16_t height,uintptr_t imageAddr,bool isWithMask);
 void ldImageFrameUpdate(ldImage_t* pWidget);
 void ldImageLoop(ldImage_t *pWidget,const arm_2d_tile_t *pParentTile,bool bIsNewFrame);
 void ldImageDel(ldImage_t *pWidget);
 
 void ldImageSetBgColor(ldImage_t *pWidget,ldColor bgColor);
 void ldImageSetOpacity(ldImage_t *pWidget, uint8_t opacity);
-void ldImageSetImage(ldImage_t *pWidget, uint32_t imageAddr, bool isWithMask);
+void ldImageSetImage(ldImage_t *pWidget, uintptr_t imageAddr, bool isWithMask);
 void ldImageSetGrayscale(ldImage_t *pWidget, uint8_t grayBit, ldColor writeColor);
 
 #define ldImageSetHidden          ldBaseSetHidden
