+++ src/gui/ldTable.h
@@ -40,8 +40,8 @@ typedef struct{
     ldColor textColor;
     ldColor itemBgColor;
     ldColor releaseImgMaskColor;
-    uint32_t releaseImgAddr;
-    uint32_t pressImgAddr;
+    uintptr_t releaseImgAddr;
+    uintptr_t pressImgAddr;
     arm_2d_region_t imgRegion;
     ldFontDict_t* pFontDict;
     uint8_t textMax;
@@ -78,8 +78,8 @@ void ldTableSetItemText(ldTable_t *pWidget,uint8_t row,uint8_t column,uint8_t *p
 void ldTableSetItemStaticText(ldTable_t *pWidget,uint8_t row,uint8_t column,uint8_t *pText,ldFontDict_t* pFontDict);
 void ldTableSetItemColor(ldTable_t *pWidget,uint8_t row,uint8_t column,ldColor textColor,ldColor bgColor);
 void ldTableSetItemAlign(ldTable_t *pWidget, uint8_t row, uint8_t column, uint8_t align);
-void ldTableSetItemImage(ldTable_t *pWidget, uint8_t row, uint8_t column, int16_t x, int16_t y, int16_t width, int16_t height, uint32_t imgAddr, bool isWithMask,ldColor maskColor,bool isMask);
-void ldTableSetItemButton(ldTable_t *pWidget,uint8_t row,uint8_t column,int16_t x,int16_t y,int16_t width,int16_t height,uint32_t releaseImgAddr,uint32_t pressImgAddr,bool isCheckable);
+void ldTableSetItemImage(ldTable_t *pWidget, uint8_t row, uint8_t column, int16_t x, int16_t y, int16_t width, int16_t height, uintptr_t imgAddr, bool isWithMask,ldColor maskColor,bool isMask);
+void ldTableSetItemButton(ldTable_t *pWidget,uint8_t row,uint8_t column,int16_t x,int16_t y,int16_t width,int16_t height,uintptr_t releaseImgAddr,uintptr_t pressImgAddr,bool isCheckable);
 
 uint8_t ldTableCurrentColumn(ldTable_t *pWidget);
 uint8_t ldTableCurrentRow(ldTable_t *pWidget);
